move LoadingBar to its own file
